+++ lib/widgets/dialogs/dialogs.dart
@@ -24,3 +24,31 @@ class ShowRemoveConfirmationDialog extends StatelessWidget {
     );
   }
 }
+
+class ShowOpenTrailerLinkDialog extends StatelessWidget {
+  final String movieName;
+  final String? trailerName;
+  const ShowOpenTrailerLinkDialog(
+      {Key? key, required this.trailerName, required this.movieName})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: Text(trailerName ?? 'No trailer name'),
+      content: Text(
+        'Opening trailer in YouTube browser or app for the movie $movieName',
+      ),
+      actions: [
+        TextButton(
+          onPressed: () => Navigator.pop(context, false),
+          child: Text('Dismiss'),
+        ),
+        TextButton(
+          onPressed: () => Navigator.pop(context, true),
+          child: Text('Let\'s go'),
+        ),
+      ],
+    );
+  }
+}