+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.3.0
 * CL-38: Show confirmation dialog when removing saved record
+* CL-46: Sharing youtube link
 * CL-48: Open trailer link in YouTube app or browser